RainEngine Rewrite (#15)
* Resource!!!!

* Clean TextureSystem and SoundSystem, Load lang outside of jar

* Update SmeagleBus

* Double deltatime

* New TileInfo class, semi simplify collisions

* BaseNPC class

* Vector2!

* Vector3 class

* CONSTANTS time

* Strubium/rain rewrite textureatlas (#14)

* TextureAtlas (weird but works)

* Base RenderableArtifact class, smaller texture atlas

* cleanup

* TextureRegion ify

* Use static Instance, improve MapInfoParser

* Fix lighting on sub maps

* move over example for the pr

* generateJavadoc method

+++ src/main/java/com/toxicrain/rainengine/core/registries/manager/NPCManager.java
@@ -22,7 +22,7 @@ public class NPCManager {
     }
 
     // Update all NPCs and their behaviors
-    public void update(float deltaTime) {
+    public void update(double deltaTime) {
         // Loop through all NPCs and update their behavior
         for (NPC npc : npcs) {
             // Execute the NPC's behavior sequence if it's defined
